refactor(orderflow): take the step-0 seam repairs as further pure moves
`code-review` at `medium` returned no correctness finding and four
cleanups on the shape of the split, all taken here as pure moves:

- `event_cap_key` and `cap_events` move from the projection root into
  `projection/model.rs`, beside the `with_live` that calls the cap, so the
  read model no longer reaches back into its parent for a private helper.
- `normalized_log_intensity` and `normalized_area_size` move into
  `projection/model.rs` too, re-exported from the root so
  `projection::normalized_*` keeps resolving for the chart's live strip;
  `tiers.rs` and `fold.rs` now import from `super::model` and depend on
  the root for nothing.
- The six bubble-only and eight lane-only inline tests move out of
  `config.rs` into `#[cfg(test)] mod tests` blocks of `config/bubbles.rs`
  and `config/lane.rs` -- a surface that moves out takes its tests with
  it -- which lets `SERIALIZED_FLOAT_PLACES` go back to private and drops
  the two imports the root tests carried for them.
- `config/lane.rs` imports the bubble items from `super::bubbles` rather
  than through the root's public re-export list.

Bodies unchanged: the statement-line multisets are still 966/966 and
544/544, and the config test code compared with the tests included is
1,189/1,189.
