+++ crates/orderflow/src/config/lane.rs
@@ -8,10 +8,10 @@
 
 use serde::{Deserialize, Serialize};
 
-use super::{
-    BubbleStyle, MAX_BUBBLE_CLUSTER_MS, MAX_BUBBLE_MAX_RADIUS, MAX_BUBBLE_MIN_RADIUS,
-    MIN_BUBBLE_MAX_RADIUS, finite_clamp,
+use super::bubbles::{
+    BubbleStyle, MAX_BUBBLE_MAX_RADIUS, MAX_BUBBLE_MIN_RADIUS, MIN_BUBBLE_MAX_RADIUS,
 };
+use super::{MAX_BUBBLE_CLUSTER_MS, finite_clamp};
 use crate::history::TapeAge;
 
 /// Default share of the chart width taken by the live lane. Enough room for
@@ -552,5 +552,480 @@ impl LiveLaneStyle {
         let min = (bubbles.min_radius * scale).clamp(0.0, MAX_BUBBLE_MIN_RADIUS);
         let max = (bubbles.max_radius * scale).clamp(MIN_BUBBLE_MAX_RADIUS, MAX_BUBBLE_MAX_RADIUS);
         (min.min(max), max)
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+    use crate::config::DEFAULT_BUBBLE_CLUSTER_MS;
+
+    /// The lane is a pane of the chart, so its width is a share of the chart
+    /// and nothing else — the candle zoom has no say in it at all.
+    #[test]
+    fn the_lane_takes_a_share_of_the_chart_and_never_more_than_half() {
+        let lane = LiveLaneStyle::default();
+        assert!((lane.resolved_width_px(1_000.0) - 350.0).abs() < 0.01);
+        assert!((lane.resolved_width_px(600.0) - 210.0).abs() < 0.01);
+
+        // A wider share takes more of the same chart; half is the ceiling.
+        let wide = LiveLaneStyle {
+            width_share: 0.9,
+            ..LiveLaneStyle::default()
+        };
+        let mut capped = wide.clone();
+        capped.sanitize();
+        assert_eq!(capped.width_share, MAX_LIVE_LANE_SHARE);
+        assert!((capped.resolved_width_px(1_000.0) - 500.0).abs() < 0.01);
+        // Even unsanitized input cannot reach past half the chart.
+        assert!(wide.resolved_width_px(1_000.0) <= 500.01);
+
+        // Degenerate charts get no lane rather than a width the layout would
+        // have to invent a meaning for.
+        assert_eq!(lane.resolved_width_px(f32::NAN), 0.0);
+        assert_eq!(lane.resolved_width_px(0.0), 0.0);
+        // A tiny share stays a band rather than a hairline...
+        let sliver = LiveLaneStyle {
+            width_share: MIN_LIVE_LANE_SHARE,
+            ..LiveLaneStyle::default()
+        };
+        assert!(sliver.resolved_width_px(200.0) >= MIN_LIVE_LANE_WIDTH_PX);
+        // ...unless the chart itself is smaller than the floor, where the cap
+        // wins and the tape still leaves history half the window.
+        assert!((sliver.resolved_width_px(20.0) - 10.0).abs() < 0.01);
+    }
+
+    /// The lane's own zoom: same band, a different amount of market time in
+    /// it. Zooming out has to aggregate as well as compress, or the crowd it
+    /// creates is unreadable.
+    #[test]
+    fn the_lane_zoom_scales_the_window_and_the_clustering_with_it() {
+        let reference = 8_000; // a typical bar of eight seconds
+        let default = LiveLaneStyle {
+            cluster_ms: Some(100),
+            ..LiveLaneStyle::default()
+        };
+        assert_eq!(default.window_ms(reference), 8_000);
+        assert_eq!(
+            default.effective_cluster_ms(DEFAULT_BUBBLE_CLUSTER_MS, reference),
+            100
+        );
+
+        // Zoomed out: four times the market time in the same band, and a
+        // cluster four times as long — so a cluster keeps its width on screen.
+        let out = LiveLaneStyle {
+            window: LaneWindow::Auto { zoom: 0.25 },
+            ..default.clone()
+        };
+        assert_eq!(out.window_ms(reference), 32_000);
+        assert_eq!(
+            out.effective_cluster_ms(DEFAULT_BUBBLE_CLUSTER_MS, reference),
+            400
+        );
+
+        // Zoomed in: a quarter of the time, a quarter of the window.
+        let into = LiveLaneStyle {
+            window: LaneWindow::Auto { zoom: 4.0 },
+            ..default.clone()
+        };
+        assert_eq!(into.window_ms(reference), 2_000);
+        assert_eq!(
+            into.effective_cluster_ms(DEFAULT_BUBBLE_CLUSTER_MS, reference),
+            25
+        );
+
+        // Raw is a choice, not a size: it survives every zoom.
+        let raw = LiveLaneStyle {
+            cluster_ms: Some(0),
+            window: LaneWindow::Auto {
+                zoom: MIN_LIVE_LANE_ZOOM,
+            },
+            ..LiveLaneStyle::default()
+        };
+        assert_eq!(
+            raw.effective_cluster_ms(DEFAULT_BUBBLE_CLUSTER_MS, reference),
+            0
+        );
+
+        // Bounds hold whatever the bars did: a session-long reference cannot
+        // turn the tape into a second chart, and a burst cannot make it an
+        // instant wide.
+        let out = LiveLaneStyle {
+            window: LaneWindow::Auto {
+                zoom: MIN_LIVE_LANE_ZOOM,
+            },
+            ..LiveLaneStyle::default()
+        };
+        assert_eq!(out.window_ms(i64::MAX), MAX_LIVE_LANE_WINDOW_MS);
+        let into = LiveLaneStyle {
+            window: LaneWindow::Auto {
+                zoom: MAX_LIVE_LANE_ZOOM,
+            },
+            ..LiveLaneStyle::default()
+        };
+        assert_eq!(into.window_ms(1), MIN_LIVE_LANE_WINDOW_MS);
+        assert_eq!(into.effective_cluster_ms(1, 8_000), 1);
+    }
+
+    /// A pinned window is a window like any other: it resolves whatever the
+    /// bars do, and the clustering scales with it exactly as the zoom's does.
+    /// Without that second half a two-minute tape draws two minutes of prints
+    /// at a tape's density, which is a smear.
+    #[test]
+    fn a_pinned_window_ignores_the_bars_and_clusters_like_the_width_it_is() {
+        let reference = 8_000; // a typical bar of eight seconds
+        let pinned = LiveLaneStyle {
+            window: LaneWindow::Fixed { ms: 120_000 },
+            cluster_ms: Some(100),
+            ..LiveLaneStyle::default()
+        };
+        // Two minutes is two minutes, whatever the bars are doing...
+        assert_eq!(pinned.window_ms(reference), 120_000);
+        assert_eq!(pinned.window_ms(1), 120_000);
+        assert_eq!(pinned.window_ms(600_000), 120_000);
+        // ...and fifteen times the automatic window gathers fifteen times the
+        // market time into each mark.
+        assert_eq!(
+            pinned.effective_cluster_ms(DEFAULT_BUBBLE_CLUSTER_MS, reference),
+            1_500
+        );
+        // Past the clustering ceiling the scaling stops rather than running
+        // away: a mark may gather a crowd, never a whole tape.
+        let widest = LiveLaneStyle {
+            window: LaneWindow::Fixed {
+                ms: MAX_LIVE_LANE_WINDOW_MS,
+            },
+            ..pinned.clone()
+        };
+        assert_eq!(
+            widest.effective_cluster_ms(DEFAULT_BUBBLE_CLUSTER_MS, reference),
+            MAX_BUBBLE_CLUSTER_MS
+        );
+        let modest = LiveLaneStyle {
+            window: LaneWindow::Fixed { ms: 16_000 },
+            ..pinned.clone()
+        };
+        assert_eq!(
+            modest.effective_cluster_ms(DEFAULT_BUBBLE_CLUSTER_MS, reference),
+            200,
+            "twice the automatic window, twice the cluster"
+        );
+        // A narrower pinned window buys detail, the same way zooming in does.
+        let narrow = LiveLaneStyle {
+            window: LaneWindow::Fixed { ms: 4_000 },
+            ..pinned.clone()
+        };
+        assert_eq!(
+            narrow.effective_cluster_ms(DEFAULT_BUBBLE_CLUSTER_MS, reference),
+            50
+        );
+        // A wider tape always gathers more than a narrower one — the property
+        // the numbers above are examples of.
+        assert!(
+            modest.effective_cluster_ms(DEFAULT_BUBBLE_CLUSTER_MS, reference)
+                > narrow.effective_cluster_ms(DEFAULT_BUBBLE_CLUSTER_MS, reference)
+        );
+        // Raw is a choice, not a size: it survives a pinned window too.
+        let raw = LiveLaneStyle {
+            cluster_ms: Some(0),
+            ..pinned
+        };
+        assert_eq!(
+            raw.effective_cluster_ms(DEFAULT_BUBBLE_CLUSTER_MS, reference),
+            0
+        );
+
+        // Absurd durations are clamped to what the tape can draw, both ends.
+        let mut wild = LiveLaneStyle {
+            window: LaneWindow::Fixed { ms: i64::MAX },
+            ..LiveLaneStyle::default()
+        };
+        wild.sanitize();
+        assert_eq!(
+            wild.window,
+            LaneWindow::Fixed {
+                ms: MAX_LIVE_LANE_WINDOW_MS
+            }
+        );
+        let mut tiny = LiveLaneStyle {
+            window: LaneWindow::Fixed { ms: 0 },
+            ..LiveLaneStyle::default()
+        };
+        tiny.sanitize();
+        assert_eq!(
+            tiny.window,
+            LaneWindow::Fixed {
+                ms: MIN_LIVE_LANE_WINDOW_MS
+            }
+        );
+    }
+
+    /// The gesture over the tape edits whichever language the window is in,
+    /// and never swaps the language underneath the trader.
+    #[test]
+    fn a_zoom_gesture_never_changes_which_mode_the_tape_is_in() {
+        let mut following = LaneWindow::default();
+        following.zoom_by(2.0);
+        assert_eq!(following, LaneWindow::Auto { zoom: 2.0 });
+
+        let mut pinned = LaneWindow::Fixed { ms: 60_000 };
+        pinned.zoom_by(2.0);
+        assert_eq!(
+            pinned,
+            LaneWindow::Fixed { ms: 30_000 },
+            "zooming in halves the pinned window; it does not return to automatic"
+        );
+        pinned.zoom_by(0.5);
+        assert_eq!(pinned, LaneWindow::Fixed { ms: 60_000 });
+
+        // A pinned window cannot be dragged out of the drawable range.
+        let mut floor = LaneWindow::Fixed {
+            ms: MIN_LIVE_LANE_WINDOW_MS,
+        };
+        floor.zoom_by(1_000.0);
+        assert_eq!(
+            floor,
+            LaneWindow::Fixed {
+                ms: MIN_LIVE_LANE_WINDOW_MS
+            }
+        );
+
+        // A meaningless gesture is refused rather than applied as garbage.
+        let mut untouched = LaneWindow::Fixed { ms: 60_000 };
+        untouched.zoom_by(f32::NAN);
+        untouched.zoom_by(0.0);
+        untouched.zoom_by(-1.0);
+        assert_eq!(untouched, LaneWindow::Fixed { ms: 60_000 });
+    }
+
+    /// An empty tape has two possible meanings and only one of them is
+    /// "nothing is trading". The axis has to separate them.
+    ///
+    /// The lane's right edge follows the newer of the book clock and the print
+    /// clock; only prints draw bubbles. So a book running ahead of the tape
+    /// puts every bubble left of the edge, and past the whole window puts none
+    /// of them on the tape — while the prints themselves are still on the
+    /// chart, in the slot of the bar they happened in. A scalper reading an
+    /// empty tape as a still market is reading a delivery gap as a fact about
+    /// the market, which is the expensive kind of wrong.
+    #[test]
+    fn the_axis_says_how_old_the_newest_mark_is() {
+        let behind = |ms| Some(TapeAge::Behind(ms));
+        // Current: nothing to say. Not "0 s behind" — a tape that is on the
+        // edge must not carry a caption about being late.
+        assert_eq!(lane_lag_label(30_000, None), None);
+        // Under the threshold: an ordinary lull, and the axis stays quiet.
+        assert_eq!(lane_lag_label(30_000, behind(3_000)), None);
+        // Visibly behind, but the newest mark is still on the tape, so the
+        // wording points at the mark the reader can find.
+        assert_eq!(
+            lane_lag_label(30_000, behind(6_000)),
+            Some("last print 6 s back".to_owned())
+        );
+        // Exactly one window behind, the mark sits on the tape's leftmost
+        // pixel — still there, so the wording must still point at it. Only
+        // past the window does the tape genuinely hold nothing.
+        assert_eq!(
+            lane_lag_label(30_000, behind(30_000)),
+            Some("last print 30 s back".to_owned())
+        );
+        assert_eq!(
+            lane_lag_label(30_000, behind(30_001)),
+            Some("no print for 30 s".to_owned())
+        );
+
+        // The floor follows the window, because "behind" is relative to what
+        // the tape is showing: four seconds is most of a four-second tape and
+        // invisible on a two-minute one.
+        assert_eq!(
+            lane_lag_label(4_000, behind(4_000)),
+            Some("last print 4 s back".to_owned())
+        );
+        assert_eq!(lane_lag_label(120_000, behind(4_000)), None);
+        assert_eq!(lane_lag_label(120_000, behind(17_000)), None);
+        assert_eq!(
+            lane_lag_label(120_000, behind(30_000)),
+            Some("last print 30 s back".to_owned())
+        );
+
+        // …but never below a second, whatever the share works out to. The
+        // tape zooms to MIN_LIVE_LANE_WINDOW_MS, where a sixth is 33 ms — less
+        // than the gap between two ordinary prints, so a share-only floor
+        // would light the caption permanently on a perfectly current tape.
+        assert_eq!(lane_lag_label(MIN_LIVE_LANE_WINDOW_MS, behind(285)), None);
+        assert_eq!(lane_lag_label(MIN_LIVE_LANE_WINDOW_MS, behind(900)), None);
+        assert_eq!(
+            lane_lag_label(MIN_LIVE_LANE_WINDOW_MS, behind(1_500)),
+            Some("no print for 1 s".to_owned()),
+            "a second and a half of silence on a 200 ms tape is a real hole"
+        );
+
+        // Nothing has printed at all. There is no mark to be late, so no
+        // window threshold applies and the wording never sends the reader
+        // looking for a bubble. This is the state every restart and every
+        // symbol switch during a quiet stretch lands in — the chart opened at
+        // lunch, which is the report this whole branch came from.
+        assert_eq!(
+            lane_lag_label(30_000, Some(TapeAge::NothingYet(4_000))),
+            Some("no print for 4 s".to_owned())
+        );
+        assert_eq!(
+            lane_lag_label(MIN_LIVE_LANE_WINDOW_MS, Some(TapeAge::NothingYet(90_000))),
+            Some("no print for 1 min 30 s".to_owned()),
+            "an empty tape is worth saying at any zoom"
+        );
+
+        // A degenerate window must not divide the label by zero or claim a
+        // current tape is late.
+        assert_eq!(lane_lag_label(0, None), None);
+        assert_eq!(
+            lane_lag_label(0, behind(5_000)),
+            Some("no print for 5 s".to_owned())
+        );
+    }
+
+    /// A menu that offers "auto" against "30 s" is asking the trader to
+    /// compare durations, so the automatic entry has to state the one it
+    /// currently amounts to.
+    #[test]
+    fn the_automatic_entry_states_the_duration_it_currently_works_out_to() {
+        let following = LaneWindow::default();
+        assert_eq!(lane_window_label(following, Some(8_000)), "auto (≈ 8 s)");
+        assert_eq!(
+            lane_window_label(following, Some(90_000)),
+            "auto (≈ 1 min 30 s)"
+        );
+        // With no bars to measure, it names the policy rather than inventing
+        // a number.
+        assert_eq!(lane_window_label(following, None), "auto");
+        // A zoomed automatic window reports what it shows, not the reference.
+        let zoomed = LaneWindow::Auto { zoom: 4.0 };
+        assert_eq!(lane_window_label(zoomed, Some(8_000)), "auto (≈ 2 s)");
+
+        for (ms, expected) in [
+            (250, "250 ms"),
+            (15_000, "15 s"),
+            (60_000, "1 min"),
+            (120_000, "2 min"),
+            (300_000, "5 min"),
+        ] {
+            assert_eq!(
+                lane_window_label(LaneWindow::Fixed { ms }, Some(8_000)),
+                expected
+            );
+        }
+        // Every preset the menu offers reads as the round number it was
+        // chosen as.
+        for ms in LANE_WINDOW_PRESETS_MS {
+            let label = lane_window_label(LaneWindow::Fixed { ms }, Some(8_000));
+            assert!(!label.contains("ms"), "{label} is not how a trader says it");
+        }
+
+        // Selecting the automatic entry while already following the bars must
+        // not reset a zoom the trader set; a preset only matches its own
+        // duration.
+        assert!(same_lane_window(zoomed, LaneWindow::default()));
+        assert!(same_lane_window(
+            LaneWindow::Fixed { ms: 60_000 },
+            LaneWindow::Fixed { ms: 60_000 }
+        ));
+        assert!(!same_lane_window(
+            LaneWindow::Fixed { ms: 60_000 },
+            LaneWindow::Fixed { ms: 30_000 }
+        ));
+        assert!(!same_lane_window(zoomed, LaneWindow::Fixed { ms: 60_000 }));
+    }
+
+    /// Every preset written before the tape had a window mode says "follow the
+    /// bars" by saying nothing, and has to keep drawing what it drew.
+    #[test]
+    fn a_file_from_before_the_mode_existed_opens_exactly_as_it_did() {
+        let old: LiveLaneStyle = toml::from_str(
+            "width_share = 0.4\ntime_zoom = 2.0\nradius_scale = 1.5\nshow_marks = false\n",
+        )
+        .unwrap();
+        assert_eq!(old.window, LaneWindow::Auto { zoom: 2.0 });
+        assert_eq!(
+            old.window_ms(8_000),
+            4_000,
+            "the same window it always drew"
+        );
+        assert!((old.width_share - 0.4).abs() < 1e-6);
+        assert!(!old.show_marks);
+        // A file that never named the tape's layers gets the tape's defaults —
+        // there is nothing to inherit from any more.
+        assert!(
+            old.enabled,
+            "a file from before the switch opens with a tape"
+        );
+        assert!(old.show_depth && old.show_aggressions);
+
+        // A file that said `false` said it on purpose, and is still obeyed —
+        // the one thing the switch to a plain `bool` may not quietly discard.
+        let chosen: LiveLaneStyle =
+            toml::from_str("show_depth = false\nshow_aggressions = false\nenabled = false\n")
+                .unwrap();
+        assert!(!chosen.show_depth, "an explicit no stays a no");
+        assert!(!chosen.show_aggressions);
+        assert!(!chosen.enabled, "and a tape put away stays away");
+
+        // A file this build writes round-trips, in both languages.
+        for window in [
+            LaneWindow::default(),
+            LaneWindow::Auto { zoom: 0.5 },
+            LaneWindow::Fixed { ms: 120_000 },
+        ] {
+            let style = LiveLaneStyle {
+                window,
+                enabled: false,
+                show_depth: false,
+                show_aggressions: true,
+                ..LiveLaneStyle::default()
+            };
+            let text = toml::to_string(&style).unwrap();
+            assert_eq!(toml::from_str::<LiveLaneStyle>(&text).unwrap(), style);
+        }
+
+        // Pinning parks the zoom at its default rather than hoarding the last
+        // one: "automatic" means the bars decide, not that a multiplier is
+        // waiting to reappear.
+        let pinned = LiveLaneStyle {
+            window: LaneWindow::Fixed { ms: 60_000 },
+            ..LiveLaneStyle::default()
+        };
+        let text = toml::to_string(&pinned).unwrap();
+        assert!(text.contains("window_ms = 60000"), "{text}");
+        assert!(text.contains("time_zoom = 1.0"), "{text}");
+    }
+
+    #[test]
+    fn the_lane_overrides_only_what_it_was_given() {
+        let bubbles = BubbleStyle::default();
+        let tuned = LiveLaneStyle {
+            cluster_ms: Some(50),
+            radius_scale: 2.0,
+            ..LiveLaneStyle::default()
+        };
+        assert_eq!(
+            tuned.effective_cluster_ms(DEFAULT_BUBBLE_CLUSTER_MS, 8_000),
+            50
+        );
+        let (min, max) = tuned.scaled_radii(&bubbles);
+        assert!((min - bubbles.min_radius * 2.0).abs() < 1e-4);
+        assert!((max - bubbles.max_radius * 2.0).abs() < 1e-4);
+        // Even an absurd scale stays inside the radius bounds the renderer
+        // was built for, and never inverts the range.
+        let huge = LiveLaneStyle {
+            radius_scale: MAX_LIVE_LANE_RADIUS_SCALE,
+            ..LiveLaneStyle::default()
+        };
+        let (min, max) = huge.scaled_radii(&BubbleStyle {
+            min_radius: MAX_BUBBLE_MIN_RADIUS,
+            max_radius: MAX_BUBBLE_MAX_RADIUS,
+            ..BubbleStyle::default()
+        });
+        assert!(min <= max);
+        assert!(max <= MAX_BUBBLE_MAX_RADIUS);
+        assert!(min <= MAX_BUBBLE_MIN_RADIUS);
     }
 }
